# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telecom Offer Advisor API",
    "description": "Phase 5: employee approval, voice and typed AI conversations, verified replies, saved outcomes, temporary speech caching, and dashboard estimates. Money uses integer minor units (100 = 1 AZN); timestamps use UTC. Accepted means confirmed interest for employee processing, never activation.",
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Dashboard",
        "description": "Phase 4: aggregate metrics from recommendations and saved call outcomes.",
        "operationId": "dashboard_api_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/recommendations": {
      "get": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Recommendations",
        "description": "Phase 2: sort by score descending, savings descending, then customer ID ascending.",
        "operationId": "recommendations_api_recommendations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 150
                },
                {
                  "type": "null"
                }
              ],
              "description": "Customer name search",
              "title": "Search"
            },
            "description": "Customer name search"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationListResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/customers/{customer_id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Customer",
        "description": "Phase 2: profile, three months of usage, recommendation, and previous call/result.",
        "operationId": "customer_api_customers__customer_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/recommendations/{recommendation_id}/approve": {
      "post": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Approve",
        "description": "Phase 3: idempotent approval; approved_by is a demo label, not authenticated identity.",
        "operationId": "approve_api_recommendations__recommendation_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Recommendation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/calls": {
      "post": {
        "tags": [
          "Calls"
        ],
        "summary": "Start Call",
        "description": "Phase 3: require approval and contact eligibility; retries reuse start_request_id.",
        "operationId": "start_call_api_calls_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartCallResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/calls/{call_id}/turns": {
      "post": {
        "tags": [
          "Calls"
        ],
        "summary": "Turn",
        "description": "Typed or recorded customer input; retries reuse client_turn_id within a call.\n\nThe browser caps recordings at 20 seconds. The server caps files at 2 MiB and the\ncomplete multipart body at 3 MiB; validates MIME/container signatures; transcribes\nEnglish speech; and sends the transcript through the same controller as typed input.\nRetry audio with the identical file, or the saved transcript as text, using the same ID.",
        "operationId": "turn_api_calls__call_id__turns_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Call Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_turn_api_calls__call_id__turns_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApproveRequest": {
        "properties": {
          "approved_by": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Approved By",
            "default": "demo_employee"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApproveRequest"
      },
...
      "CallStatus": {
        "type": "string",
        "enum": [
          "active",
          "completed",
          "unresolved"
        ],
        "title": "CallStatus"
      },
...
      "DashboardResponse": {
        "properties": {
          "recommended_customers": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Recommended Customers"
          },
          "potential_monthly_savings_minor": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Potential Monthly Savings Minor"
          },
          "currency": {
            "type": "string",
            "const": "AZN",
            "title": "Currency",
            "default": "AZN"
          },
          "calls_started": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Calls Started"
          },
          "calls_completed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Calls Completed"
          },
          "call_completion_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Call Completion Rate"
          },
          "accepted_offers": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Accepted Offers"
          },
          "acceptance_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Acceptance Rate"
          },
          "accepted_monthly_savings_minor": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Accepted Monthly Savings Minor"
          },
          "estimated_staff_minutes_saved": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Estimated Staff Minutes Saved"
          },
          "staff_time_assumption": {
            "type": "string",
            "title": "Staff Time Assumption",
            "default": "5 minutes manual handling minus 1 minute employee review"
          },
          "outcome_counts": {
            "$ref": "#/components/schemas/OutcomeCounts"
          },
          "top_opportunities": {
            "items": {
              "$ref": "#/components/schemas/RecommendationResponse"
            },
            "type": "array",
            "title": "Top Opportunities"
          },
          "resettable_demo_data": {
            "type": "boolean",
            "const": true,
            "title": "Resettable Demo Data",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "recommended_customers",
          "potential_monthly_savings_minor",
          "calls_started",
          "calls_completed",
          "call_completion_rate",
          "accepted_offers",
          "acceptance_rate",
          "accepted_monthly_savings_minor",
          "estimated_staff_minutes_saved",
          "outcome_counts",
          "top_opportunities"
        ],
        "title": "DashboardResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "RecommendationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "RecommendationListResponse"
      },
...
      "StartCallRequest": {
        "properties": {
          "recommendation_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Recommendation Id"
          },
          "start_request_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Start Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "recommendation_id",
          "start_request_id"
        ],
        "title": "StartCallRequest"
      },
      "StartCallResponse": {
        "properties": {
          "call_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Call Id"
          },
          "status": {
            "$ref": "#/components/schemas/CallStatus"
          },
          "conversation_state": {
            "$ref": "#/components/schemas/ConversationState"
          },
          "introduction_turn": {
            "$ref": "#/components/schemas/CallTurnResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "call_id",
          "status",
          "conversation_state",
          "introduction_turn"
        ],
        "title": "StartCallResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Shared demo access token; never an OpenAI key.",
        "scheme": "bearer"
      }
    }
  }
}